+++ .gitignore
@@ -5,12 +5,12 @@
 *.pdf
 *.psd
 *.sublime*
-build/css
-build/fonts
-build/img
-build/leaflet
-build/favicon.ico
-build/index.html
+#build/css
+#build/fonts
+#build/img
+#build/leaflet
+#build/favicon.ico
+#build/index.html
 node_modules/
 npm-debug.log*
 Thumbs.db
