+++ source/js/error.js
@@ -1,5 +1,9 @@
-import {allFieldset, allMapFilters, mainForm, mapFeatures, mapFilterForm, pageBody} from './utils.js';
+import {mainForm, mapFilterForm, pageBody} from './utils.js';
 import {addDisabledValue} from './helpers.js';
+
+const mapFeatures = document.querySelector('.map__features');
+const allMapFilters = mapFilterForm.querySelectorAll('.map__filter');
+const allFieldset = mainForm.querySelectorAll('fieldset');
 
 const errorResponse = err => {
