+++ source/js/use-form.js
@@ -1,25 +1,16 @@
-import {
-  initCoords,
-  mainForm,
-  mainPinMarker,
-  mapPoints,
-  mapFilterForm,
-  typeFormInput,
-  priceInput,
-  timeinSelect,
-  timeoutSelect,
-  roomNumberSelect,
-  capacitySelect
-} from './utils.js';
-import { setAddressValue, getMinPrice, getRoomsValue } from './helpers.js';
+import {priceInput, mainForm} from './utils.js';
+import {getMinPrice, getRoomsValue} from './helpers.js';
 import uploadImage from './upload-image.js';
 
 
-const useForm = () => {
-  const btnReset = mainForm.querySelector('.ad-form__reset');
-  const avatar = mainForm.querySelector('.ad-form-header__preview > img');
-  const photo = mainForm.querySelector('.ad-form__photo');
+const typeFormInput = mainForm.querySelector('#type');
+const timeinSelect = mainForm.querySelector('#timein');
+const timeoutSelect = mainForm.querySelector('#timeout');
+const roomNumberSelect = mainForm.querySelector('#room_number');
+const capacitySelect = mainForm.querySelector('#capacity');
 
+
+const useForm = () => {
 
   typeFormInput.addEventListener('change' , (evt) => {
     const currentValue = evt.target.value;
@@ -51,10 +42,10 @@ const useForm = () => {
   priceInput.addEventListener('input', (evt) => {
 
     const inputValue = evt.target.value;
-    const LIMIT = getMinPrice(typeFormInput.value);
+    const priceLimit = getMinPrice(typeFormInput.value);
 
-    if (inputValue < LIMIT) {
-      evt.target.setCustomValidity(`Цена должна не менее чем ${LIMIT} руб.`);
+    if (inputValue < priceLimit) {
+      evt.target.setCustomValidity(`Цена должна не менее чем ${priceLimit} руб.`);
     }
     else {
       evt.target.setCustomValidity('');
@@ -66,16 +57,7 @@ const useForm = () => {
   uploadImage('#avatar', '.ad-form-header__preview > img');
   uploadImage('#images', '.ad-form__photo', true);
 
-  btnReset.addEventListener('click', ()=> {
-    mainForm.reset();
-    priceInput.placeholder = '1000';
-    mapFilterForm.reset();
-    avatar.src = 'img/muffin-grey.svg';
-    photo.style.backgroundImage = '';
-    mapPoints.clearLayers();
-    setAddressValue();
-    mainPinMarker.setLatLng(initCoords);
-  })
+
 
 }
 export default useForm;
